LPS-88392 Move related fake service classes into corresponding *Test as inner class
