+++ portal-impl/test/integration/com/liferay/portal/kernel/lar/StagedModelDataHandlerRegistryUtilTest.java
@@ -14,9 +14,11 @@
 
 package com.liferay.portal.kernel.lar;
 
+import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.exportimport.kernel.lar.StagedModelDataHandler;
 import com.liferay.exportimport.kernel.lar.StagedModelDataHandlerRegistryUtil;
-import com.liferay.portal.kernel.lar.bundle.stagedmodeldatahandlerregistryutil.TestStagedModelDataHandler;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.registry.Registry;
 import com.liferay.registry.RegistryUtil;
@@ -24,6 +26,7 @@ import com.liferay.registry.ServiceRegistration;
 
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.junit.AfterClass;
 import org.junit.Assert;
@@ -91,5 +94,115 @@ public class StagedModelDataHandlerRegistryUtilTest {
 
 	private static ServiceRegistration<StagedModelDataHandler>
 		_serviceRegistration;
+
+	private static class TestStagedModelDataHandler
+		implements StagedModelDataHandler<User> {
+
+		public static final String[] CLASS_NAMES = {
+			TestStagedModelDataHandler.class.getName()
+		};
+
+		@Override
+		public void deleteStagedModel(
+			String uuid, long groupId, String className, String extraData) {
+		}
+
+		@Override
+		public void deleteStagedModel(User user) {
+		}
+
+		@Override
+		public void exportStagedModel(
+			PortletDataContext portletDataContext, User user) {
+		}
+
+		@Override
+		public User fetchMissingReference(String uuid, long groupId) {
+			return null;
+		}
+
+		@Override
+		public User fetchStagedModelByUuidAndGroupId(
+			String uuid, long groupId) {
+
+			return null;
+		}
+
+		@Override
+		public List<User> fetchStagedModelsByUuidAndCompanyId(
+			String uuid, long companyId) {
+
+			return null;
+		}
+
+		@Override
+		public String[] getClassNames() {
+			return CLASS_NAMES;
+		}
+
+		@Override
+		public String getDisplayName(User user) {
+			return null;
+		}
+
+		@Override
+		public int[] getExportableStatuses() {
+			return null;
+		}
+
+		@Override
+		public Map<String, String> getReferenceAttributes(
+			PortletDataContext portletDataContext, User user) {
+
+			return null;
+		}
+
+		/**
+		 * @deprecated As of Wilberforce (7.0.x)
+		 */
+		@Deprecated
+		@Override
+		public void importCompanyStagedModel(
+			PortletDataContext portletDataContext, Element element) {
+		}
+
+		/**
+		 * @deprecated As of Wilberforce (7.0.x)
+		 */
+		@Deprecated
+		@Override
+		public void importCompanyStagedModel(
+			PortletDataContext portletDataContext, String uuid, long classPK) {
+		}
+
+		@Override
+		public void importMissingReference(
+			PortletDataContext portletDataContext, Element referenceElement) {
+		}
+
+		@Override
+		public void importMissingReference(
+			PortletDataContext portletDataContext, String uuid, long groupId,
+			long classPK) {
+		}
+
+		@Override
+		public void importStagedModel(
+			PortletDataContext portletDataContext, User user) {
+		}
+
+		@Override
+		public void restoreStagedModel(
+			PortletDataContext portletDataContext, User user) {
+		}
+
+		@Override
+		public boolean validateReference(
+			PortletDataContext portletDataContext, Element referenceElement) {
+
+			return false;
+		}
+
+	}
 
 }
